bugfix:  CarbonDataLoadingException: Error while initializing data handler :

+++ flinkx-carbondata/flinkx-carbondata-writer/src/main/java/com/dtstack/flinkx/carbondata/writer/TaskNumberGenerator.java
@@ -11,8 +11,8 @@ public class TaskNumberGenerator {
      * Generate unique number to be used as partition number of file name
      */
     public static String generateUniqueNumber(int taskId, String segmentId, long partitionNumber) {
-        return String.valueOf((int)(Math.pow(10,2)) + Integer.valueOf(segmentId))
-                + String.valueOf((int)Math.pow(10,5) + taskId)
-                + String.valueOf(partitionNumber + (int)Math.pow(10,5));
+        return String.valueOf((long)(Math.pow(10,2)) + Integer.valueOf(segmentId))
+                + String.valueOf((long)Math.pow(10,5) + taskId)
+                + String.valueOf(partitionNumber + (long)Math.pow(10,5));
     }
 }
